pyproject.toml
[tool.poetry]
name = "dyn-cloudflare-ddns"
version = "0.1.1"
description = "Update Cloudflare DNS from a dynamic DNS record"
authors = ["Guillermo Castro <admin@codegeek.dev>"]
packages = [
    { include = "dyn-cloudflare-ddns" }
]

[tool.poetry.dependencies]
python = "^3.10"
cloudflare = "^2.19.2"
dnspython = "^2.1.0"
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.1.7"]
build-backend = "poetry.core.masonry.api"
